refactor: consolidate external food search logic into dedicated service
